rename NW-UCLA dataset file name
