+++ torch_skeleton/datasets/__init__.py
@@ -3,4 +3,4 @@ __all__ = ["DiskCache", "Apply", "NTU", "UCLA"]
 from .base_dataset import DiskCache, Apply
 
 from .ntu import NTU
-from .nw_ucla import UCLA
+from .ucla import UCLA
